requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.31.0
pydantic==2.11.4
pydantic-settings==2.9.0
httpx==0.28.1
teler==0.2.0
numpy==1.26.4
scipy==1.15.3
google-genai==1.36.0
